Fixed score remains same after going to prev question
Added a new property to each of the questions in quizData,
'correctAnswer', which should have a value of either true or false.

Changed the logic in the checkAnswer function to change the
'correctAnswer' to true/false whether the selected answer is
correct/incorrect.

Then added a new function calculateScore which goes through
quizData at the end of the quiz and adds +1 to score if correctAnswer
is true for each of the questions. This function gets called in endQuiz.